refactor(redux): simplify redux process
